Adds an optional dark mode to the site
Dropped a bunch of `black` tachyons classes to make this work; also
switch to solarized-dark since I can't seem to control the background
color of the `pre` blocks due to the way the highlighter works.

This also changes the way the underlines on the links get created
because the previous way was not working across lines.

Closes #9.