Add style guide
via https://www.bryanbraun.com/styleguide/ and
https://poormansstyleguide.com. The markdown version is enough for me,
because this site is static and has no form elements.

This page should be left permanently in draft as it is only useful for
getting a sense of how the site's CSS displays. (Although maybe some
day, if my CSS is great, I'll actually want to display it? Who knows!).