css: Make anchors less intrusive
Currently anchors are always showing in the same font as the
headers. This changes the anchors so that they aren't visible until
hovering over the header itself, at which point they'll show in a light
grey.

It's possible this won't work all that well for mobile, so may end up
modifying at least the visibility approach.