rss: Try fixing by not minifying
When deploying, we turn on hugo's minification - which is usually good!
But for the RSS feed, it seems like it's removing all the leading spaces
in code blocks, which isn't so good. Disable minification for xml (rss
feeds) but enable it for everything else, by default.

Found this fix here:
https://jdheyburn.co.uk/blog/who-goes-blogging-7-hugo-minify-rss-code-indentation-fix/