chore: Fix stylelint pre-commit-config
I changed this previously to match [this article][1], but in the
article, the dependencies for the hook are defined in the pre-commit
hook. Instead, I want the versions to match the ones in package.json,
which will be picked up by my other tooling (null-ls in neovim,
specifically).

It seems that if I _don't_ specify the versions here, this doesn't work,
so switch back to using the system language and shelling out to `npx`.

[1]: https://adamj.eu/tech/2023/02/09/pre-commit-hooks-unsupported-tools/#node-hook-stylelint-a-css-linter