meta: Trim down dependencies
I could never quite get these to work how I'd like them to, and they're
a bit more of a hindrance right now. I'd like to go more minimal on the
dependencies.

The only one of these I remember being a real use was stylelint (which
_was_ useful for my limited CSS knowledge), but I don't want to keep a
dependency around that needs to be updated as often as a typical NPM
package. Plus, being on sr.ht, I don't have something like dependabot to
automate this (and I wouldn't _want_ half of my commits to be dependabot
updates).

For now, rip off this bandaid. I'm open to adding small, light
dependencies in the future.