now: Update "Last updated" and move things around
I hadn't updated this in awhile - tried to figure out if I could
"automatically" do it with hugo, but didn't have much luck figuring out
how to make the single page template. Instead, just update the <time>
manually and add a comment reminding me to update it in the future.