css: Change to nested
This lets the various large-screen styles be closer to their
small-screen counterparts.

One thing that is unfortunate is that I can't set a variable to reuse
for all of the "min-width: 45em" bits here. Tried this, and ran into
[this StackOverflow answer][1]. I imagine there's probably some other
cleaner way to do this, but I'm not thinking of it right now.

Ref: https://css-irl.info/css-nesting-is-here/

[1]: https://stackoverflow.com/questions/40722882/css-native-variables-not-working-in-media-queries